Moved fragments to frontend root
