+++ src/components/OptionRequestForm.tsx
@@ -9,9 +9,9 @@ import { OptionFlavor } from "../models/option";
 import { ethers } from "ethers";
 import { loadNFTImage, imageOf, keyOf } from "../datasources/NFT/localhost";
 import { showToast } from "../utils/frontend/toasting";
-import TextBox_OptionRequestForm from "../fragments/TextBox.OptionRequest";
+import TextBox_OptionRequestForm from "../frontend/fragments/TextBox.OptionRequest";
 import Radio_OptionRequestForm from "../fragments/Radio.OptionRequest";
-import DropDown_OptionRequestForm from "../fragments/DropDown.OptionRequest";
+import DropDown_OptionRequestForm from "../frontend/fragments/DropDown.OptionRequest";
 import { useContracts } from "../pages/_app";
 import { OptionRequest } from "../models/optionRequest";
 import { NFTAsset } from "../models/nftAsset";
